feat(examples): all examples are now E2E regression tests
