+++ examples/consumer/goconsumer/client.go
@@ -44,7 +44,7 @@ func (c *Client) login(username string, password string) (*User, error) {
       "password": "%s"
     }`, username, password)
 
-	res, err := http.Post(fmt.Sprintf("%s/users/login", c.Host), "application/json", bytes.NewReader([]byte(loginRequest)))
+	res, err := http.Post(fmt.Sprintf("%s/users/login", c.Host), "application/json; charset=utf-8", bytes.NewReader([]byte(loginRequest)))
 	if res.StatusCode != 200 || err != nil {
 		return nil, fmt.Errorf("login failed")
 	}
