+++ examples/consumer/goconsumer/user_service_test.go
@@ -7,10 +7,12 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 
 	"github.com/pact-foundation/pact-go/dsl"
+	"github.com/pact-foundation/pact-go/types"
 )
 
 // Common test data
@@ -18,10 +20,20 @@ var dir, _ = os.Getwd()
 var pactDir = fmt.Sprintf("%s/../../pacts", dir)
 var logDir = fmt.Sprintf("%s/log", dir)
 var pact dsl.Pact
-var loginRequest = ` { "username":"billy", "password": "issilly" }`
 var form url.Values
 var rr http.ResponseWriter
 var req *http.Request
+
+// var name = "Jean-Marie de La Beaujardière😀😍"
+var name = "billy"
+var like = dsl.Like
+var eachLike = dsl.EachLike
+var term = dsl.Term
+var loginRequest = fmt.Sprintf(`{ "username":"%s", "password": "issilly" }`, name)
+
+var commonHeaders = map[string]string{
+	"Content-Type": "application/json; charset=utf-8",
+}
 
 // Use this to control the setup and teardown of Pact
 func TestMain(m *testing.M) {
@@ -82,7 +94,16 @@ func TestPactConsumerLoginHandler_UserExists(t *testing.T) {
 
 		return nil
 	}
+	body :=
+		like(fmt.Sprintf(
+			`{
+            "user": {
+              "name": "%s",
+              "type": %v
+            }
+					}`, name, term("admin", "admin|user|guest")))
 
+	// Pull from pact broker, used in e2e/integrated tests for pact-go release
 	// Setup interactions on the Mock Service. Note that you can have multiple
 	// interactions
 	pact.
@@ -96,16 +117,7 @@ func TestPactConsumerLoginHandler_UserExists(t *testing.T) {
 		}).
 		WillRespondWith(dsl.Response{
 			Status: 200,
-			Headers: map[string]string{
-				"Content-Type": "application/json; charset=utf-8",
-			},
-			Body: `
-				{
-				  "user": {
-				    "name": "billy"
-				  }
-				}
-			`,
+			Body:   body,
 		})
 
 	err := pact.Verify(testBillyExists)
@@ -133,15 +145,14 @@ func TestPactConsumerLoginHandler_UserDoesNotExist(t *testing.T) {
 		Given("User billy does not exist").
 		UponReceiving("A request to login with user 'billy'").
 		WithRequest(dsl.Request{
-			Method: "POST",
-			Path:   "/users/login",
-			Body:   loginRequest,
+			Method:  "POST",
+			Path:    "/users/login",
+			Body:    loginRequest,
+			Headers: commonHeaders,
 		}).
 		WillRespondWith(dsl.Response{
-			Status: 404,
-			Headers: map[string]string{
-				"Content-Type": "application/json; charset=utf-8",
-			},
+			Status:  404,
+			Headers: commonHeaders,
 		})
 
 	err := pact.Verify(testBillyDoesNotExists)
@@ -169,19 +180,42 @@ func TestPactConsumerLoginHandler_UserUnauthorised(t *testing.T) {
 		Given("User billy is unauthorized").
 		UponReceiving("A request to login with user 'billy'").
 		WithRequest(dsl.Request{
-			Method: "POST",
-			Path:   "/users/login",
-			Body:   loginRequest,
+			Method:  "POST",
+			Path:    "/users/login",
+			Body:    loginRequest,
+			Headers: commonHeaders,
 		}).
 		WillRespondWith(dsl.Response{
-			Status: 401,
-			Headers: map[string]string{
-				"Content-Type": "application/json; charset=utf-8",
-			},
+			Status:  401,
+			Headers: commonHeaders,
 		})
 
 	err := pact.Verify(testBillyUnauthorized)
 	if err != nil {
 		t.Fatalf("Error on Verify: %v", err)
+	}
+}
+
+func TestPactConsumerLoginHandler_Publish(t *testing.T) {
+	// Enable when running E2E/integration tests before a release
+	if os.Getenv("PACT_INTEGRATED_TESTS") != "" {
+		var brokerHost = os.Getenv("PACT_BROKER_HOST")
+
+		// Publish the Pacts...
+		p := dsl.Publisher{}
+		err := p.Publish(types.PublishRequest{
+			PactURLs:        []string{filepath.ToSlash(fmt.Sprintf("%s/billy-bobby.json", pactDir))},
+			PactBroker:      brokerHost,
+			ConsumerVersion: "1.0.0",
+			Tags:            []string{"latest", "sit4"},
+			BrokerUsername:  os.Getenv("PACT_BROKER_USERNAME"),
+			BrokerPassword:  os.Getenv("PACT_BROKER_PASSWORD"),
+		})
+
+		if err != nil {
+			t.Fatal("Error:", err)
+		}
+	} else {
+		t.Skip()
 	}
 }
