+++ examples/mux/provider/user_service_test.go
@@ -33,6 +33,61 @@ func TestPact_Provider(t *testing.T) {
 	if err != nil {
 		t.Fatal("Error:", err)
 	}
+
+	// Pull from pact broker, used in e2e/integrated tests for pact-go release
+	if os.Getenv("PACT_INTEGRATED_TESTS") != "" {
+		var brokerHost = os.Getenv("PACT_BROKER_HOST")
+
+		// Verify the Provider - Specific Published Pacts
+		err = pact.VerifyProvider(types.VerifyRequest{
+			ProviderBaseURL:            fmt.Sprintf("http://127.0.0.1:%d", port),
+			PactURLs:                   []string{fmt.Sprintf("%s/pacts/provider/bobby/consumer/billy/latest/sit4", brokerHost)},
+			ProviderStatesSetupURL:     fmt.Sprintf("http://127.0.0.1:%d/setup", port),
+			BrokerUsername:             os.Getenv("PACT_BROKER_USERNAME"),
+			BrokerPassword:             os.Getenv("PACT_BROKER_PASSWORD"),
+			PublishVerificationResults: true,
+			ProviderVersion:            "1.0.0",
+			Verbose:                    true,
+		})
+
+		if err != nil {
+			t.Fatal("Error:", err)
+		}
+
+		// Verify the Provider - Latest Published Pacts for any known consumers
+		err = pact.VerifyProvider(types.VerifyRequest{
+			ProviderBaseURL:            fmt.Sprintf("http://127.0.0.1:%d", port),
+			BrokerURL:                  brokerHost,
+			ProviderStatesSetupURL:     fmt.Sprintf("http://127.0.0.1:%d/setup", port),
+			BrokerUsername:             os.Getenv("PACT_BROKER_USERNAME"),
+			BrokerPassword:             os.Getenv("PACT_BROKER_PASSWORD"),
+			PublishVerificationResults: true,
+			ProviderVersion:            "1.0.0",
+			Verbose:                    true,
+		})
+
+		if err != nil {
+			t.Fatal("Error:", err)
+		}
+
+		// Verify the Provider - Tag-based Published Pacts for any known consumers
+		err = pact.VerifyProvider(types.VerifyRequest{
+			ProviderBaseURL:            fmt.Sprintf("http://127.0.0.1:%d", port),
+			ProviderStatesSetupURL:     fmt.Sprintf("http://127.0.0.1:%d/setup", port),
+			BrokerURL:                  brokerHost,
+			Tags:                       []string{"latest", "sit4"},
+			BrokerUsername:             os.Getenv("PACT_BROKER_USERNAME"),
+			BrokerPassword:             os.Getenv("PACT_BROKER_PASSWORD"),
+			PublishVerificationResults: true,
+			ProviderVersion:            "1.0.0",
+		})
+
+		if err != nil {
+			t.Fatal("Error:", err)
+		}
+	} else {
+		t.Log("Skipping pulling from broker as PACT_INTEGRATED_TESTS is not set")
+	}
 }
 
 // Starts the provider API with hooks for provider states.
@@ -95,6 +150,7 @@ var billyExists = &examples.UserRepository{
 			Name:     "billy",
 			Username: "billy",
 			Password: "issilly",
+			Type:     "admin",
 		},
 	},
 }
@@ -107,6 +163,7 @@ var billyUnauthorized = &examples.UserRepository{
 			Name:     "billy",
 			Username: "billy",
 			Password: "issilly1",
+			Type:     "blocked",
 		},
 	},
 }
